ipfunctions: Adding upstream "Help File"
The "help" for the functions is actually just a Google Doc which hasn't
been edited since 2013.

This is an HTML rendering of the Google Doc which _*may*_ be able to be
referenced in the IFRAME sandbox allowed by AppScript.

Regardless, in the event that it cannot this should be considered a
reference.